+++ core/src/org/ode4j/ode/internal/joints/DxJoint.java
@@ -796,7 +796,7 @@ public abstract class DxJoint extends DObject implements DJoint, Cloneable {
 
 				// the angle we get will be between 0..2*pi, but we want to return angles
 				// between -pi..pi
-				if ( theta > M_PI ) theta -= ( double )( 2 * M_PI );
+				if ( theta > M_PI ) theta -= 2 * M_PI;
 
 				// the angle we've just extracted has the wrong sign
 				theta = -theta;
